pyproject.toml
[project]
name = "model-m2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "streamlit>=1.50.0",
    "ultralytics>=8.3.209",
]
